fn.higherorder.HigherOrder moved to adt.intlist.IntList (aligned with slides)
