+++ src/main/scala/adt/intlist/IntList.scala
@@ -1,13 +1,27 @@
 package adt.intlist
 
+
+sealed trait IntList {
+  def exists(func: Int => Boolean): Boolean =
+    ???
+
+  def filter(func: Int => Boolean): IntList =
+    ???
+
+  // def find(func: Int => Boolean): ??? =
+  //   ???
+}
+
+final case class IntPair(head: Int, tail: IntList)
+  extends IntList
+
+case object IntNil extends IntList
+
 object Main extends App {
-  // val ints = IntPair(1, IntPair(2, IntPair(3, IntNil())))
+  val ints = IntPair(1, IntPair(3, IntPair(5, IntNil)))
 
-  // println(s"""$ints.contains(1) == ${ints.contains(1)}""")
-  // println(s"""$ints.contains(5) == ${ints.contains(5)}""")
-
-  // println(s"""$ints.add(1) == ${ints.add(1)}""")
-  // println(s"""$ints.add(5) == ${ints.add(5)}""")
-
-  // println(s"""$ints.total == ${ints.total}""")
+  println(s"""ints.exists(_ > 0)      == ${ints.exists(_ > 0)}""")
+  println(s"""ints.exists(_ < 0)      == ${ints.exists(_ < 0)}""")
+  println(s"""ints.exists(_ % 2 == 0) == ${ints.exists(_ % 2 == 0)}""")
+  println(s"""ints.exists(_ % 2 == 1) == ${ints.exists(_ % 2 == 1)}""")
 }
